Changed version of Jackson using bom 2.7.9 (see description for the reason)
SpringBoot 2.1.6 uses Jacskon 2.9.9, but the deserialization performed by Jackson seems to be afflicted by a bug (https://stackoverflow.com/questions/47701937/what-version-of-jackson-does-spring-boot-use, see also https://github.com/FasterXML/jackson-databind/issues/1599). Within the versions fixed so far, there is the 2.7.9. The problem happened in the deserialization of the followedUsers list of a user in the database, as happened with the test testAddFollowedUser_ShouldBeUpdatedWithSuccess()